requirements.txt
typer[All]
SoccerNet
matplotlib
scikit-learn
pytorch-lightning
opencv-python
moviepy
tqdm
tabulate
timm
mmengine